Integrate front end html,js with backend php #72
Minimum repository version stands for the first submission, we get the
first submission of each user and the Id of that record to do sql query.
The due date is marked 0. One day after is marked +1. One day before is
marked -1.Also add those days with 0 submission, these days can now be
displayed on chart.