Fix logic error in two php #73
I found that first submission time distribution chart and last submission
time distributionchart are the same, which means logic error in php. In the old code, I use "select max(RepositoryVersion) rv,FKUserId,Id group by
FKUserId" to get Id of first/last submission. This is wrong because Id
will always be the same regardless of first or last submission. In the new
approach, I get the Id of first/last submission in another query. So the
logic error in the two php is fixed.