Implement the logic in php #92
One option is to display mark by GPA (F, P, C, D, HD). The second option
is to display mark at a 10% step (0-10%, 10-20%,20-30%...). First get the
max mark of all submissions of an assignment for each user. Then increase
the count that each mark belongs to according to the option that user
chooses.