Implement studentActivitiesOverview front end #123
Implement the chart "Student Activities Overview". Implement related
functions: change presentation order, set start day and end day, threshold
function, Implement the datepicker to enable the user to choose start day and
end day in the calendar. Implement the mechanism to deal with x-label
issue: to ensure the accuracy and correctness of information, it will
dynamically determine which stage we are at when the left position and
right position of the slider are changed: at the first stage, display no
x-label (e-charts display some of the labels by default because there is
not enough space to accommodate all the x-labels, this will provide
misleading information); at the second stage, display all x-labels at a 90
degree; at the third stage, display all x-labels at a 40 degree; at the
last stage, display all x-labels normally (at a 0 degree).