Implement web page StudentActivitiesDetails #123
The purpose of this page is to show "Event name", "Event context" and the
corresponding amount of the user (given the course and the period). The
data will be displayed in two tables. Also provide hints and hyper links
to the two tables in case the first table is too lengthy for the user to
become aware of the existence of the second table